+++ android/src/main/java/com/pichillilorenzo/flutter_inappwebview/service_worker/ServiceWorkerChannelDelegate.java
@@ -1,5 +1,6 @@
 package com.pichillilorenzo.flutter_inappwebview.service_worker;
 
+import android.annotation.SuppressLint;
 import android.os.Build;
 import android.webkit.ServiceWorkerController;
 
@@ -35,6 +36,7 @@ public class ServiceWorkerChannelDelegate extends ChannelDelegateImpl {
     this.serviceWorkerManager = serviceWorkerManager;
   }
 
+  @SuppressLint("RestrictedApi")
   @Override
   public void onMethodCall(@NonNull MethodCall call, @NonNull MethodChannel.Result result) {
     ServiceWorkerControllerCompat serviceWorkerController = ServiceWorkerManager.serviceWorkerController;
@@ -72,6 +74,13 @@ public class ServiceWorkerChannelDelegate extends ChannelDelegateImpl {
           result.success(false);
         }
         break;
+      case "getRequestedWithHeaderMode":
+        if (serviceWorkerWebSettings != null && WebViewFeature.isFeatureSupported(WebViewFeature.REQUESTED_WITH_HEADER_CONTROL)) {
+          result.success(serviceWorkerWebSettings.getRequestedWithHeaderMode());
+        } else {
+          result.success(null);
+        }
+        break;
       case "getCacheMode":
         if (serviceWorkerWebSettings != null && WebViewFeature.isFeatureSupported(WebViewFeature.SERVICE_WORKER_CACHE_MODE)) {
           result.success(serviceWorkerWebSettings.getCacheMode());
@@ -104,6 +113,13 @@ public class ServiceWorkerChannelDelegate extends ChannelDelegateImpl {
         if (serviceWorkerWebSettings != null && WebViewFeature.isFeatureSupported(WebViewFeature.SERVICE_WORKER_CACHE_MODE)) {
           Integer mode = (Integer) call.argument("mode");
           serviceWorkerWebSettings.setCacheMode(mode);
+        }
+        result.success(true);
+        break;
+      case "setRequestedWithHeaderMode":
+        if (serviceWorkerWebSettings != null && WebViewFeature.isFeatureSupported(WebViewFeature.REQUESTED_WITH_HEADER_CONTROL)) {
+          Integer mode = (Integer) call.argument("mode");
+          serviceWorkerWebSettings.setRequestedWithHeaderMode(mode);
         }
         result.success(true);
         break;